requirements.txt
torch
datasets
transformers
peft
trl
bitsandbytes
faiss-cpu 
sentence-transformers
tiktoken
